feat(devx-soar): Report bot raised PRs once week
We have a few bots raising PRs on a regular basis,
and we're struggling to triage them.

This can make it difficult to identify the non-bot PRs (human PRs) in the list,
meaning those PRs can stay open for multiple days/weeks.

In an attempt to reduce noise, this change creates two workflows to notify:
  - Of bot and human PRs every Monday
  - Of human PRs Tuesday - Friday

The hope is we'll be more likely to triage the bot PRs if we see them less often.