Add janus to prnouncer for DevX/Sec
The internal Janus repository gets a lot of PRs that aren't the
responsibility of DevX (access changes). However, we would still like
to be told about changes that come from our automation tooling.

This PR adds another PRnouncer step to DevX/Sec's workflow that only
announces PRs from our automation bots. The only repository that this
applies to for now is Janus, so I've simplified the configuration by
inlining the `github-repositories` configuration.

I've included more bots then Janus actually uses, so that this is easy
to extend to other repos in the future, even if they have different
automations in place.